pyproject.toml
[project]
name = "pytest-loco"
version = "1.3.0"
description = "Another one YAML-based DSL for testing"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Plugins",
    "Framework :: Pytest",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Topic :: Utilities",
]
authors = [
    {name = "Mikhalev Oleg", email = "mhalairt@gmail.com"}
]
license = "BSD-2-Clause"
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">3.13,<4"
dependencies = [
    "click (>=8.3.1,<9.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
]

[project.urls]
"Source Code" = "https://github.com/pytest-loco/pytest-loco"
"Issues" = "https://github.com/pytest-loco/pytest-loco/issues"

[project.entry-points.pytest11]
pytest_loco = "pytest_loco.plugin"

[project.scripts]
pytest-loco = "pytest_loco.__main__:cli"

[tool.poetry]
packages = [{include = "pytest_loco", from = "src"}]

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
mypy = {extras = ["mypyc", "reports", "stubgen"], version = "^1.19.1"}
pre-commit = "^4.5.1"
ruff = ">=0.14.11,<0.15.0"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pyfakefs = "^6.0.0"
pytest = "^9.0.2"
pytest-cov = "^7.0.0"
pytest-mock = "^3.15.1"
pytest-xdist = "^3.8.0"

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.format]
docstring-code-format = true
quote-style = "single"

[tool.ruff.lint]
select = [
    "ANN",     # flake8-annotations
    "ASYNC",   # flake8-async
    "S",       # flake8-bandit
    "A",       # flake8-builtins
    "B",       # flake8-bugbear
    "COM",     # flake8-commas
    "C4",      # flake8-comprehensions
    "T20",     # flake8-print
    "Q",       # flake8-quotes
    "RSE",     # flake8-raise
    "SLF",     # flake8-self
    "SIM",     # flake8-simplify
    "TC",      # flake8-type-checking
    "ARG",     # flake8-unused-arguments
    "I",       # isort
    "E", "W",  # pycodestyle
    "D",       # pydocstyle
    "F",       # pyflakes
    "PL",      # pylint
    "UP",      # pyupgrade
    "RUF",     # ruff
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["ANN", "S"]

[tool.ruff.lint.isort]
known-first-party = ["pytest_loco"]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"
multiline-quotes = "single"

[tool.ruff.lint.flake8-type-checking]
quote-annotations = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.13"
plugins = [
    "pydantic.mypy",
]
mypy_path = ["src"]
packages = ["pytest_loco"]
strict = true
exclude_gitignore = true
explicit_package_bases = true
pretty = true
show_error_codes = true
color_output = true

[[tool.mypy.overrides]]
module = "pytest_loco.__main__"
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_any_generics = false
disallow_untyped_decorators = false
disallow_untyped_defs = false
disallow_incomplete_defs = false
warn_return_any = false
allow_redefinition = true
ignore_errors = true

[[tool.mypy.overrides]]
module = "yaml.*"
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
minversion = "9.0.0"
testpaths = ["tests"]
addopts = [
    "-q",
    "-n", "auto",
    "-p no:pytest_loco",
    "--cov=pytest_loco",
    "--cov-report=term",
    "--cov-report=html",
    "--strict-markers",
    "--strict-config",
]

[tool.coverage.run]
source = ["pytest_loco"]
branch = true
parallel = true
concurrency = ["multiprocessing"]

[tool.coverage.paths]
source = [
    "src/pytest_loco",
    "**/site-packages/pytest_loco"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if typing.TYPE_CHECKING:",
    "if TYPE_CHECKING:",
]
fail_under = 80
show_missing = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
